### 2024-08-01 #209 @jsdodge
THzTools submission
Submitting Author: (@jsdodge)
Package Name: THzTools
Data analysis software tools for terahertz time-domain spectroscopy (THz-TDS)
Repository Link:  https://github.com/dodge-research-group/thztools
Version submitted: 0.5.4
Editor:@banesullivan
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10100093.svg)](https://doi.org/10.5281/zenodo.10100093)
Version accepted: 0.5.5

---

*THzTools* provides tools to simplify and improve procedures for data analysis in terahertz time-domain spectroscopy (THz-TDS). Some of the methods included in the package were described previously in the paper at this [link](https://doi.org/10.1364/OE.417724). As the name suggests, terahertz time-domain spectroscopy involves measurements of terahertz-frequency electromagnetic waveforms that are are acquired as a function of *time*. A variety of methods exist to transform these measurements into functions of *frequency*, but the standard procedures have several pitfalls. *THzTools* provides software tools that make it easier for researchers to use the best available methods for analyzing their data.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is researchers working with THz-TDS, although the procedures may be useful in other areas that use time-domain measurement systems. The package is designed for characterizing the time-domain noise performance of THz-TDS measurement systems and for analyzing the results from these systems in the frequency domain.

The [Fit-TDS](https://github.com/THzbiophotonics/Fit-TDS) package provides a graphical user interface that simplifies THz-TDS data analysis with standard analysis methods. THzTools focuses on lower-level statistical procedures, and implements algorithms that are not available in Fit-TDS.

  - If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or `@tag` the editor you contacted: @NickleDave

- [x] The package is deposited in a long-term repository with the DOI: https://zenodo.org/doi/10.5281/zenodo.10100093
