### 2024-08-01 04a2c21
Update README to reflect change from alpha to beta development status
