+++ README.md
@@ -13,7 +13,7 @@ represent as a function of *frequency* for analysis. THzTools makes it easier fo
 to use statistically optimal methods for doing this analysis, as described in [L. Mohtashemi et al.,
  Opt. Express **29**, 4912 (2021)](https://doi.org/10.1364/OE.417724).
 
-This is *alpha* software that is currently under development.
+This is *beta* software that is currently under development.
 
 | Information       | Links                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                |
 |:------------------|:-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
