### 2024-08-13 #209 @cmarmo
Hello @jsdodge ! Thank you for submitting THzTools to pyOpenSci.
Sorry for the delay of my answer!
I'm Chiara and I'm going to take care of your submission for the initial editorial checks.
I will be back to you by the end of the week.
Thanks for your patience!
