### 2024-08-15 #209 @cmarmo
> Do you have any suggestions for ways that we can continue development without disrupting the review? Would it be ok, for example, to continue developing on the dev branch as long as we don't merge these into the main branch?

I'm starting the editorial checks right now: I can perform them on the main branch if you prefer, once we agree on a version to be submitted you can tag the last modifications and we can update the description of the issue. Would that be ok for you?
