### 2024-08-15 #209 @cmarmo
Hi @jsdodge ! Thank you again for submitting your package for pyOpenSci review.
Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

---
---

THzTools is in excellent shape, congratulations!
I have some minor comments before starting looking for an editor:
- in the `README.md` file installation and development instructions are missing: a simple `pip install` and/or `conda install` will be enough (this explicits the fact that the package is available from both the channels), and then links to installation and development pages in the documentation.
- the `CONTRIBUTING.md` file is missing: do you mind adding it with some basic instructions and a link to the contribuuting section in the documentation?
- I noticed that in the issue template for bug reports the explanation says  "help us improve SciPy" ... :grin:

> >  Do you have any suggestions for ways that we can continue development without disrupting the review? Would it be ok, for example, to continue developing on the dev branch as long as we don't merge these into the main branch?
>
>
> I'm starting the editorial checks right now: I can perform them on the main branch if you prefer, once we agree on a version to be submitted you can tag the last modifications and we can update the description of the issue. Would that be ok for you?

I realize my answer was a bit out of scope... sorry for that. Indeed, I believe it is a good idea to continue the development in a separate branch during review: however, reviewers might ask for modifications too and everything would in principle end in a new version accepted at the end of the review process. Please just clarify with reviewers in which branch you are addressing their comments, we had some misunderstanding in the past.
