### 2024-08-15 #209 @jsdodge
Hi Chiara,

Thanks! We can add installation instructions to the `README.md` file and change the issue template right away. This raises a similar question to my earlier one: should we do this in the `main` branch? Normally we would also bump the version number when making changes in `main`, but the submission version is listed as v0.5.0.

Regarding the `CONTRIBUTING.md` file, we have a [`contributing.rst`](https://github.com/dodge-research-group/thztools/blob/main/docs/source/contributing.rst) file in `/docs/source/`. Could you recommend a way to include this information at the top level in a `CONTRIBUTING.md` file *and* in the documentation without duplicating it? A related question is whether we can do the same thing with the `CODE_OF_CONDUCT.md` file. Currently we just have a GitHub link to `CODE_OF_CONDUCT.md` in `contributing.rst`, but it would be better to include it directly in the documentation. We used a link because we didn't know how to use ReST to pull it into the documentation source.
