### 2024-08-15 #209 @cmarmo
> Thanks! We can add installation instructions to the README.md file and change the issue template right away. This raises a similar question to my earlier one: should we do this in the main branch? Normally we would also bump the version number when making changes in main, but the submission version is listed as v0.5.0.

Technically the review is not started yet: once done with the changes we can edit the issue description.

> Regarding the CONTRIBUTING.md file, we have a [contributing.rst](https://github.com/dodge-research-group/thztools/blob/main/docs/source/contributing.rst) file in /docs/source/. Could you recommend a way to include this information at the top level in a CONTRIBUTING.md file and in the documentation without duplicating it?

Some general information would be enough in the `CONTRIBUTING.md` file: a link to `contributing.rst` there will complete the instructions. See for example what is done in one of the previously [accepted packages](https://github.com/sourmash-bio/sourmash/blob/latest/CONTRIBUTING.md).

> A related question is whether we can do the same thing with the CODE_OF_CONDUCT.md file. Currently we just have a GitHub link to CODE_OF_CONDUCT.md in contributing.rst, but it would be better to include it directly in the documentation. We used a link because we didn't know how to use ReST to pull it into the documentation source.

Your solution is fine with our standards: if you prefer to have the code of conduct in the documentation then you can use the same approach suggested for the `CONTRIBUTING.md` file and link the documentation reference inside.
