### 2024-08-15 cdb9c8b
Add CONTRIBUTING.md; add installation instructions to README.md; fix typo in bug_report issue template; bump version
