+++ paper/paper.bib
@@ -1,50 +1,3 @@
-@book{bechhoefer2021,
-  title = {Control {{Theory}} for {{Physicists}}},
-  author = {Bechhoefer, John},
-  year = {2021},
-  publisher = {Cambridge University Press},
-  address = {Cambridge},
-  doi = {10.1017/9780511734809}
-}
-
-@inproceedings{denakpo2022,
-  title = {Noise Analysis, Noise Reduction and Covariance Estimation for {{Time}} Domain {{Spectroscopy}}},
-  booktitle = {2022 47th {{International Conference}} on {{Infrared}}, {{Millimeter}} and {{Terahertz Waves}} ({{IRMMW-THz}})},
-  author = {Denakpo, Elsa and Hannotte, Th{\'e}o and Lavancier, M{\'e}lanie and Barois, Sophie Eliet and Peretti, Romain},
-  year = {2022},
-  pages = {1--2},
-  doi = {10.1109/IRMMW-THz50927.2022.9895812}
-}
-
-@misc{dodge2021,
-  title = {{{MATLAB}} Functions for Maximum-Likelihood Parameter Estimation in Terahertz Time-Domain Spectroscopy},
-  author = {Dodge, J. Steven and Mohtashemi, Laleh and Westlund, Paul and Sahota, Derek G. and Lea, Graham B. and Bushfield, Ian and Mousavi, Payam},
-  year = {2021},
-  doi = {10.5281/zenodo.4326594},
-  howpublished = {Zenodo}
-}
-
-@inproceedings{lavancier2021,
-  title = {A Criterion to Compare Permittivity Models in the Terahertz Range},
-  booktitle = {2021 46th {{International Conference}} on {{Infrared}}, {{Millimeter}} and {{Terahertz Waves}} ({{IRMMW-THz}})},
-  author = {Lavancier, Melanie and Yassine, Nabil Vindas and Vlieghe, Juliette and Lampin, Jean-Fran{\c c}ois and Peretti, Romain},
-  year = {2021},
-  pages = {1--2},
-  doi = {10.1109/IRMMW-THz50926.2021.9567242}
-}
-
-@article{lee2023,
-  title = {The {{dotTHz Project}}: {{A Standard Data Format}} for {{Terahertz Time-Domain Data}}},
-  shorttitle = {The {{dotTHz Project}}},
-  author = {Lee, Jongmin and Leung, Chi Ki and Ma, Mingrui and {Ward-Berry}, Jasper and Santitewagun, Supawan and Zeitler, J. Axel},
-  year = {2023},
-  journal = {Journal of Infrared, Millimeter, and Terahertz Waves},
-  volume = {44},
-  number = {11},
-  pages = {795--813},
-  doi = {10.1007/s10762-023-00947-w}
-}
-
 @article{mohtashemi2021,
   title = {Maximum-likelihood parameter estimation in terahertz time-domain spectroscopy},
   author = {Mohtashemi, Laleh and Westlund, Paul and Sahota, Derek G. and Lea, Graham B. and Bushfield, Ian and Mousavi, Payam and Dodge, J. Steven},
@@ -57,13 +10,17 @@
   doi = {10.1364/OE.417724}
 }
 
-@inproceedings{muller2021,
-  title = {Beyond the "{{Dynamic}} Range" Approach {{In}} Noise Evaluation for {{Terahertz Time}} Domain Spectrometers},
-  booktitle = {2021 46th {{International Conference}} on {{Infrared}}, {{Millimeter}} and {{Terahertz Waves}} ({{IRMMW-THz}})},
[…]
