+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+We welcome contributions to THzTools. We use [GitHub Flow](https://docs.github.com/en/get-started/using-github/github-flow)
+to manage our open-source development. Please set up a pull request against the
+`main` branch with any changes that you would like us to consider. See the
+[Contributing](https://dodge-research-group.github.io/thztools/contributing.html)
+section of the documentation for more details.
