### 2024-08-15 #209 @jsdodge
I just released [THzTools v0.5.1](https://github.com/dodge-research-group/thztools/releases/tag/v0.5.1) with the changes that you requested. I've also updated the version number in the submission documentation.
