### 2024-09-10 9eb4bc4
Fix documentation error in fit; update README; update bib-file for paper; update IDE settings
