### 2024-09-11 #209 @jsdodge
Hi @cmarmo, I have just updated the [submission](https://github.com/pyOpenSci/software-submission/issues/209#issue-2442260242) with the latest release, [v0.5.2](https://github.com/dodge-research-group/thztools/releases/tag/v0.5.2).

Could you please update me on the editor search?
