### 2024-09-12 #209 @cmarmo
Thank you @jsdodge for the follow-up

> Could you please update me on the editor search?

I'm sorry to say that I am still looking... I guess the end of summer plus the beginning of the academic year are not making things easier.... thanks for your patience!
