### 2024-10-08 #209 @lwasser
@jsdodge this is great. I'll leave a few notes and then will let @banesullivan step in. We may have a second reviewer. In the meantime, does Romain have a GitHub handle so we can add them to this issue? The review will happen fully in this issue with links to any issues or pr's opened of course!
