### 2024-10-09 #209 @lwasser
ok fantastic.
✅ we have two reviewers now assigned to this issue.  @frank1010111 and @Romain-Peretti
I believe we are at [this step](https://www.pyopensci.org/software-peer-review/how-to/editors-guide.html#respond-to-the-submitter-in-the-github-issue) in the review process and  @banesullivan can take the review forward from here!! Bane can you please onboard the reviewers and kick things off?

Thank you all. Please let me know if i can be helpful moving forward!
