### 2024-10-10 #209 @banesullivan
:wave: Hi  @frank1010111 and @Romain-Peretti! Thank you for volunteering to review THzTools for pyOpenSci! I'm excited to kick off this review and I'll try to chime in with a few comments of my own but overall THzTools is looking in great schape -- wonderful work, @jsdodge!

- [ ] @frank1010111 survey completed.
- [ ] @Romain-Peretti survey completed.

Please get in touch with any questions or concerns!

### Your review is due: November ~8th, 2024

I've set the due date ~3 weeks out from now, but we're flexible here so please don't rush yourself if life or other things pop up and require you to delay a bit.

---

Reviewers: @frank1010111 @Romain-Peretti
Due date: 2024-11-08
