### 2024-10-12 dodge-research-group/thztools#58 @frank1010111
DOC: add repostatus badge
### Issue with current documentation:

To meet the README file requirements for the ongoing [PyOpenSci review](https://github.com/pyOpenSci/software-submission/issues/209), please add a repostatus badge

### Idea or request for content:

Add a [repostatus](https://www.repostatus.org/) badge to the readme, probably this one:
[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)

### Additional context (e.g. screenshots, GIFs)

_No response_
