### 2024-10-12 20f3ee1
Update documentation for linting with Hatch
