### 2024-10-12 a59e113
Update documentation to include instructions for type checking
