### 2024-10-12 #209 @jsdodge
Hi @frank1010111 , thanks for your feedback so far. How would you like to proceed with the code updates? I've already addressed the issues that you posted, but just in my local `dev` branch. At what point would you recommend that I push these changes to the remote, merge them with the `main` branch, and create a new release? Should I wait for you to complete your review before doing any or all of these?
