### 2024-10-13 f67c651
Revise JOSS paper; update changelog
