### 2024-10-13 #209 @jsdodge
In that case, I merged the changes into the `main` branch and created a new release, [THzTools v0.5.3 (beta)](https://github.com/dodge-research-group/thztools/releases/tag/v0.5.3). I also closed the issues that you raised. I see that there are several unchecked items, so I'll wait for further recommendations.
