### 2024-10-14 #209 @banesullivan
Hey @jsdodge, I anticipate that a few more changes will be made during this review process, so there is no need to issue a new release for each of these changes during the review (unless you prefer to do so!). As the review finalizes though, we can make sure a release is issued and update the version specifier in the submission.

Moving forward, would you please try to link back to this GitHub issue by pasting the URL link in the descriptions of any Pull Requests or new Issues? This will help us track changes to the software that were a direct result of this review and help us check off items in the reviewers comments. Thanks!
