### 2024-10-16 #209 @frank1010111
Today, I checked the paper draft [created by a github action](https://github.com/dodge-research-group/thztools/actions/runs/11310914645/artifacts/2049501236) and found it meets JOSS specifications.
