### 2024-10-17 dodge-research-group/thztools#61 @frank1010111
DOC: README requirements for PyOpenSci
### Issue with current documentation:

The [review for PyOpenSci](https://github.com/pyOpenSci/software-submission/issues/209) sets out a few extra requirements for the README. This includes
* Descriptive links for all vignettes
* A package demonstration
* how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.
* Citation information

### Idea or request for content:

I would suggest adding three or four things to the readme

- [x] links to two or three of [these examples](https://dodge-research-group.github.io/thztools/examples.html)
- [x] A worked through simple example (could be one of those examples you've already got)
- [x] Elaborating on the comparison to other THz analysis tools that you mentioned in [the paper's statement of need](https://github.com/dodge-research-group/thztools/blob/main/paper/paper.md#statement-of-need) in the readme
- [x] An easy-to-copy citation in a code block near the end

For bonus points, it might be worth your while to include a [CITATION.cff](https://citation-file-format.github.io/) to the root of the project, though that's going above and beyond the requirements for the review

### Additional context (e.g. screenshots, GIFs)

_No response_
