### 2024-10-17 dodge-research-group/thztools#61 @jsdodge
Thanks for those suggestions, @frank1010111 . To clarify, since we expect to have the software published in JOSS at the end of the review, is there a provisional citation that you would recommend before that?
