### 2024-10-17 dodge-research-group/thztools#61 @frank1010111
Yes, since JOSS will require it anyway, I would suggest a following [this example to get a Zenodo](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) citation.
