### 2024-10-17 01265a0
Add CITATION.cff; update email for Santiago in pyproject.toml; update bib-file
