+++ paper/paper.bib
@@ -78,15 +78,6 @@
   doi = {10.1063/1.5047659}
 }
 
-@inproceedings{osseiran2023,
-  title = {From {{Noise}} Analysis to {{Error Bars}} on {{Refractive Index}} in {{THz-TDS}}},
-  booktitle = {2023 48th {{International Conference}} on {{Infrared}}, {{Millimeter}}, and {{Terahertz Waves}} ({{IRMMW-THz}})},
-  author = {Osseiran, Noureddin and Bichon, Jeyan and Raj, Aditya and Hannotte, Th{\'e}o and Eliet, Sophie and Peretti, Romain},
-  year = {2023},
-  pages = {1--2},
-  doi = {10.1109/IRMMW-THz57677.2023.10299169}
-}
-
 @article{peretti2019,
   title = {{{THz-TDS Time-Trace Analysis}} for the {{Extraction}} of {{Material}} and {{Metamaterial Parameters}}},
   author = {Peretti, Romain and Mitryukovskiy, Sergey and Froberger, Kevin and Mebarki, Mohamed Aniss and Eliet, Sophie and Vanwolleghem, Mathias and Lampin, Jean-Francois},
