+++ CITATION.cff
@@ -1,0 +1,73 @@
+# This CITATION.cff file was generated with cffinit.
+# Visit https://bit.ly/cffinit to generate yours today!
+
+cff-version: 1.2.0
+title: THzTools
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Jonathan
+    family-names: Posada Loaiza
+    orcid: 'https://orcid.org/0009-0004-4443-6161'
+    affiliation: Simon Fraser University
+    email: jonathan.posada1@udea.edu.co
+  - given-names: Santiago
+    family-names: Higuera-Quintero
+    email: santiago.higuera@itp3.uni-stuttgart.de
+    affiliation: Simon Fraser University
+    orcid: 'https://orcid.org/0000-0001-6905-4662'
+  - given-names: Alireza
+    family-names: Noori
+    orcid: 'https://orcid.org/0009-0009-1091-0802'
+    email: alireza_noori@sfu.ca
+    affiliation: Simon Fraser University
+  - email: lalehmo@gmail.com
+    given-names: Laleh
+    family-names: Mohtashemi
+    orcid: 'https://orcid.org/0000-0002-2706-857X'
+    affiliation: Simon Fraser University
+  - given-names: R. P.
+    family-names: Hall
+    orcid: 'https://orcid.org/0009-0004-1137-3651'
+    email: rph4@sfu.ca
+    affiliation: Simon Fraser University
+  - email: nayimam@gmail.com
+    given-names: Naod Ayalew
+    family-names: Yimam
+    orcid: 'https://orcid.org/0009-0007-7522-5027'
+    affiliation: Simon Fraser University
+  - given-names: J. Steven
+    family-names: Dodge
+    email: jsdodge@sfu.ca
+    affiliation: Simon Fraser University
+    orcid: 'https://orcid.org/0000-0003-3219-7314'
+identifiers:
+  - type: doi
+    value: 10.5281/zenodo.13924460
+    description: Zenodo
+repository-code: 'https://github.com/dodge-research-group/thztools'
+abstract: >-
+  THzTools is an open-source Python package for data
+  analysis in terahertz time-domain spectroscopy
+
+  (THz-TDS). As the name suggests, THz-TDS involves
+  measurements of terahertz-frequency
+
+  electromagnetic waveforms that are acquired as a function
+  of time, which users typically
+
+  represent as a function of frequency for analysis.
+  THzTools makes it easier for researchers
+
+  to use statistically optimal methods for doing this
+  analysis, as described in L. Mohtashemi et al.,
+   Opt. Express 29, 4912 (2021) (https://doi.org/10.1364/OE.417724).
+keywords:
+  - Time-domain terahertz spectroscopy
+  - Maximum-likelihood analysis
+license: MIT
+commit: 0661124821ece1e65b01f7e20a0e37f0c64ec825
+version: 0.5.3
+date-released: '2024-10-13'
