### 2024-10-17 c796ff0
Remove problem line break in author list
