### 2024-10-17 6e2418e
Link to examples; provide more detail about rationale; include example code; include citations
