### 2024-10-18 #209 @frank1010111
A nit for your documentation is that you've got docstrings for the parameters of your dataclasses, but Sphinx expects docstrings for your attributes. For instance, the lines [here in NoiseModel](https://github.com/dodge-research-group/thztools/blob/74945426ddd1ffe617ce7edceef18b4da3a4491e/src/thztools/thztools.py#L211-L223) are rendered fine, but Sphinx doesn't know to put them with [these attributes](https://github.com/dodge-research-group/thztools/blob/74945426ddd1ffe617ce7edceef18b4da3a4491e/src/thztools/thztools.py#L273-L276).

The solution to this oddity is something like this
```
@dataclass
class Snake:
    """Oh no, it's a snake!"""

    width: float
    """Width of the snake."""

    length: float
    """Length of the snake."""
```

Once again, it's a nit that you don't have to change. Your documentation is perfectly readable as is. Just a fun fact for you if you're as obsessive about these things as me.
