### 2024-10-18 #209 @jsdodge
Thanks for that suggestion. Just to clarify, I'm using the [`numpydoc`](https://numpydoc.readthedocs.io/en/latest/index.html) extension, so are you recommending that I repeat the class [Parameters](https://numpydoc.readthedocs.io/en/latest/format.html#parameters) section in a separate [Attributes](https://numpydoc.readthedocs.io/en/latest/format.html#documenting-classes) section, as in the `ExampleError` class listed in the Sphinx [Example NumPy Style Python Docstrings](https://www.sphinx-doc.org/en/master/usage/extensions/example_numpy.html#example-numpy)?
