+++ pyproject.toml
@@ -28,11 +28,11 @@ classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Science/Research",
   	"Programming Language :: Python",
-  	"Programming Language :: Python :: 3.8",
   	"Programming Language :: Python :: 3.9",
   	"Programming Language :: Python :: 3.10",
   	"Programming Language :: Python :: 3.11",
   	"Programming Language :: Python :: 3.12",
+  	"Programming Language :: Python :: 3.13",
   	"Programming Language :: Python :: Implementation :: CPython",
   	"Programming Language :: Python :: Implementation :: PyPy",
     "License :: OSI Approved :: MIT License",
@@ -83,10 +83,10 @@ doctest = "pytest -v --doctest-modules {args:src/thztools}"
 make-logo = ["python logo/make_logo.py; mv thztools_logo.svg docs/source/_static/;mv thztools_logo_dark.svg docs/source/_static/"]
 
 [[tool.hatch.envs.all.matrix]]
-python = ["3.8", "3.9", "3.10", "3.11", "3.12",]
+python = ["3.9", "3.10", "3.11", "3.12", "3.13", ]
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 79
 exclude = ["examples/figgen/**"]
 
