### 2024-10-19 5571610
Update CI to drop Python 3.8 and add Python 3.13
