+++ .github/workflows/pytest-with-coverage.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [ '3.10', '3.11', '3.12' ]
+        python-version: [ '3.9', '3.10', '3.11', '3.12', '3.13' ]
     runs-on: 'ubuntu-latest'
     steps:
       - name: Checkout repository
