### 2024-10-19 dodge-research-group/thztools#63 @frank1010111
DOC: Extra docstrings
### Issue with current documentation:

This is part of the [PyOpenSci review](https://github.com/pyOpenSci/software-submission/issues/209). I swear this will be the last issue relating to the docs. Once again, I'd like to thank you for writing such good and detailed documentation. Reading through the code is a breeze and checking out the examples is super easy.

Two places are missing docstrings:

- [ ] The first line of `thztools.py` ought to have a small explainer for the file contents, purpose, and major classes and functions.
- [ ] The [CommonNLL class definition](https://github.com/dodge-research-group/thztools/blob/74945426ddd1ffe617ce7edceef18b4da3a4491e/src/thztools/thztools.py#L1009)

### Idea or request for content:

I see `CommonNLL` is being used in the error fit. Just a little information about its purpose would be useful.

### Additional context (e.g. screenshots, GIFs)

_No response_
