### 2024-10-19 #209 @frank1010111
I've run and played around with the example notebooks, and I'm overall very pleased with the package and documentation. Very nice work. After that last issue is closed, my review ought to be complete and I can recommend acceptance.
