### 2024-10-19 51f0728
Add module docstring
