### 2024-10-19 a282358
Add Attribtues section for dataclasses; add docstring for internal CommonNLL dataclass
