+++ src/thztools/thztools.py
@@ -270,6 +270,20 @@ class NoiseModel:
         :attr:`dt` and ``thztools.get_option("sampling_time")`` are ``None``,
         the :attr:`sigma_tau` parameter is given in units of the sampling time.
 
+    Attributes
+    ----------
+    sigma_alpha : float
+        Additive noise amplitude.
+    sigma_beta : float
+        Multiplicative noise amplitude.
+    sigma_tau : float
+        Timebase noise amplitude.
+    dt : float or None, optional
+        Sampling time, normally in picoseconds. Default is None, which sets
+        the sampling time to ``thztools.options.sampling_time``. If both
+        :attr:`dt` and ``thztools.get_option("sampling_time")`` are ``None``,
+        the :attr:`sigma_tau` parameter is given in units of the sampling time.
+
     Warns
     -----
     UserWarning
@@ -569,6 +583,35 @@ class NoiseResult:
     Dataclass for the output of :func:`noisefit`.
 
     Parameters
+    ----------
+    noise_model : NoiseModel
+        Noise parameters, represented as a :class:`NoiseModel` object.
+    mu : ndarray, shape (n,)
+        Signal vector.
+    a : ndarray, shape (m,)
+        Signal amplitude drift vector.
+    eta : ndarray, shape (m,)
+        Signal delay drift vector.
+    fval : float
+        Value of optimized NLL cost function.
+    hess_inv : ndarray
+        Inverse Hessian matrix of optimized NLL cost function.
+    err_sigma_alpha, err_sigma_beta, err_sigma_tau : float
+        Estimated uncertainty in the noise model parameters. Set equal to 0.0
+        when the parameter is fixed.
+    err_mu : ndarray
+        Estimated uncertainty in ``mu``.
+    err_a : ndarray
+        Estimated uncertainty in ``a``.
+    err_eta : ndarray
+        Estimated uncertainty in ``eta``.
+    diagnostic : scipy.optimize.OptimizeResult
+        Instance of :class:`scipy.optimize.OptimizeResult` returned by
+        :func:`scipy.optimize.minimize`. Note that the attributes ``fun``,
+        ``jac``, and ``hess_inv`` represent functions over the internally
+        scaled parameters.
+
+    Attributes
     ----------
     noise_model : NoiseModel
         Noise parameters, represented as a :class:`NoiseModel` object.
@@ -1055,6 +1098,34 @@ def scaleshift(
 
 @dataclass
 class CommonNLL:
+    r"""
+    Dataclass for passing intermediate cost function variables.
+
+    The functions `_nll_noisefit`, `_jac_noisefit`, and `_hess_noisefit`
+    involve computations with a common set of array variables for a given
+    input. The `_nll_common` function encapsulates this computation and
+    uses the `CommonNLL` dataclass to organize the output. A future release may
+    implement caching to improve speed.
+
+    Attributes
+    ----------
+    ressq: ndarray with shape (m, n)
+        Square of residuals, equal to (x - zeta)**2.
+    vtot: ndarray with shape (m, n)
+        Estimate for time-domain variance, corrected for drift in amplitude
+        and time delay.
+    zeta: ndarray with shape (m, n)
+        Time-domain waveform estimates, corrected for drift in amplitude and
+        time delay.
+    dzeta: ndarray with shape (m, n)
+        Time-domain derivative of zeta.
+    a: ndarray with shape (m,)
+        Amplitude drift parameters.
+    zeta_f: ndarray with shape (m, n_f)
+        Fourier amplitudes of zeta.
+    exp_iweta: ndarray with shape (m, n_f)
+        Frequency response function array used to correct for time-delay drift.
+    """
     ressq: NDArray[np.float64]
[…]
