### 2024-10-19 86ca878
Reformat with Ruff
