### 2024-10-19 4c60921
Fix bug in module docstring example
