### 2024-10-19 #209 @jsdodge
Thanks for your help, @frank1010111 ! I've made the changes that you requested and created a new release, [THzTools v0.5.4 (beta)](https://github.com/dodge-research-group/thztools/releases/tag/v0.5.4), in anticipation of your review approval. I appreciate your suggestions and your promptness in completing the review.
