### 2024-10-19 #209 @jsdodge
Thanks, @Romain-Peretti , I look forward to your feedback. As noted in the [comment above](https://github.com/pyOpenSci/software-submission/issues/209#issuecomment-2424181181), I have issued a new release in response to several helpful suggestions from @frank1010111 . I've updated the [submission](https://github.com/pyOpenSci/software-submission/issues/209#issue-2442260242) to indicate the updated release.
