+++ pyproject.toml
@@ -9,7 +9,7 @@ description = """
 Data analysis software tools for terahertz time-domain spectroscopy (THz-TDS)
 """
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = "MIT"
 keywords = ["terahertz", "THz", "time-domain spectroscopy", "data analysis"]
 authors = [
