### 2024-10-23 #209 @Romain-Peretti
So I feel very the same way as @frank1010111 it is very impressive and I bet useful for quite a lot of people (at least a great playground for my next students to understand a bit all of that).

I have a question about the way the Hessian are computed. We struggled during month to find a good way to calculate the hessian in one of our project and I am not sure to follow how it is done here. It looks like you are computing the Hessian by yourself or am I wrong ?

Romain
