### 2024-10-23 #209 @jsdodge
Thank you for your feedback! I'm glad you think the software will be useful.

To your question, for the [`noisefit`](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/src/thztools/thztools.py#L1976) function we use [`scipy.minimize`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.minimize.html) with the [BFGS](https://docs.scipy.org/doc/scipy/reference/optimize.minimize-bfgs.html) method to do the optimization, which does not require the Hessian. We tried other methods that did require the Hessian, however, so we have an internal function, [`_hess_noisefit`](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/src/thztools/thztools.py#L1437) to compute it that we left in the main code, in case we find a way to use it later to improve the computational speed. We compute the Jacobian explicitly in [`_jac_noisefit`](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/src/thztools/thztools.py#L1271).

For the [`fit`](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/src/thztools/thztools.py#L2906) function, we use [`scipy.least_squares`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.least_squares.html) to do the minimization (see [here](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/src/thztools/thztools.py#L3361)), which also just uses the Jacobian. Internally, we fit for `mu` in addition to the fit parameters of the frequency response function, so the Jacobian that we compute internally different from the Jacobian with respect to the fit parameters supplied to `fit`. We compute that Jacobian [here](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/src/thztools/thztools.py#L3267).
