### 2024-11-05 #209 @Romain-Peretti
Ok sorry for the silence ... I had a bit to do outside of the office. I finished the review process the only very small comment I would have and that is maybe a misunderstanding from my side is that the author list for  JOSS is maybe the one of the Software Package but maybe adding it to the paper.md together with the affiliation would clarify it.

- [X] **A short summary** desribing the high-level functionality of the software

---
