### 2024-11-05 #209 @jsdodge
Thanks @Romain-Peretti ! I'd like to clarify your comment: the author list for the paper (linked [here](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/paper/paper.md?plain=1#L12C1-L38C19)) is the same as the authors listed in the [`pyproject.toml`](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/pyproject.toml#L15C1-L24C1) metadata (linked [here](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/pyproject.toml#L15C1-L24C1)). Are you recommending that we change that?
