### 2024-11-19 #209 @banesullivan
Hi @jsdodge, this has come together fantastically! It looks like both @frank1010111  and @Romain-Peretti have checked everything off and are recommending approval (such a quick turn around is a first for me 🎉 ). @frank1010111 and @Romain-Peretti, thank you for your time and effort on these thoughtful reviews!

*Aside: apologies for my delay!! Life happened and before I knew it 2 weeks went by. So sorry about that!*

A few quick questions before I post the review acceptance and update the metadata in the original post at the top of this thread:

1. The only item left unchecked was in @Romain-Peretti's review (follows). I'm curious if this is something you wish/hope to address or not? With hyper specific scientific tooling like this, my impression is that the available tooling is limited so I understand if there isn't much to compare it to.
    > If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.
2. Is version `0.5.4` the version you'd like to be associated with release or are there any developments from this review that have not yet been released?
3. Similarly, do you have a Zendo version and DOI for this as well?

Otherwise, I'm ready to accept THzTools and I'll start double checking everything for the JOSS automatic submission you've opted for.

Excellent work @jsdodge!!
