### 2024-11-19 68a7bc8
Add related projects to README; update CHANGELOG; bump version
