+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 Please follow the guidelines at [Common Changelog](https://common-changelog.org)
 for maintaining this file.
 
+## [0.5.5] - 2024-11-19
+
+### Added
+- Add list of related software to README
+
 ## [0.5.4] - 2024-10-19
 
 ### Added
@@ -287,6 +292,7 @@ _Changelog introduced._
 
 - Remove `tdtf`
 
+[0.5.5]: https://github.com/dodge-research-group/thztools/releases/tag/v0.5.5
 [0.5.4]: https://github.com/dodge-research-group/thztools/releases/tag/v0.5.4
 [0.5.3]: https://github.com/dodge-research-group/thztools/releases/tag/v0.5.3
 [0.5.2]: https://github.com/dodge-research-group/thztools/releases/tag/v0.5.2
@@ -301,4 +307,3 @@ _Changelog introduced._
 [0.3.3]: https://github.com/dodge-research-group/thztools/releases/tag/v0.3.3
 [0.3.2]: https://github.com/dodge-research-group/thztools/releases/tag/v0.3.2
 [0.3.1]: https://github.com/dodge-research-group/thztools/releases/tag/v0.3.1
-[0.3.0]: https://github.com/dodge-research-group/thztools/releases/tag/v0.3.0
