### 2024-11-20 #209 @jsdodge
Thank you, @banesullivan !

In response to your first question, we included the following list of related projects in the [README](https://github.com/dodge-research-group/thztools/blob/07de6ca0b786ee74c9f138e41bf81a66a4020e0c/README.md).
- https://github.com/dotTHzTAG/CaTSper
- https://github.com/THzbiophotonics/Fit-TDS
- https://github.com/YaleTHz/nelly
- https://github.com/puls-lab/phoeniks

All of these projects except Phoeniks have publications associated with them, which we cite in the JOSS [paper](https://github.com/dodge-research-group/thztools/blob/92d074f85dfe093e11225c92770c382259643834/paper/paper.md) in the same order as they appear above:

> To support this mode of analysis, the `THzTools` package provides functionality and documentation that are unavailable in existing THz-TDS analysis software ([Lee et al., 2023](https://doi.org/10.1007/s10762-023-00947-w); [Peretti et al., 2019](https://doi.org/10.1109/TTHZ.2018.2889227); [Tayveh et al., 2021](https://doi.org/10.1021/acs.analchem.1c02132)).

In response to your other questions,
2. I just bumped the version to `0.5.5`, so please use that.
3. The Zenodo DOI is [10.5281/zenodo.13955913](https://doi.org/10.5281/zenodo.10100093).
