### 2024-11-23 #209 @banesullivan
---

🎉 With that, THzTools has been approved by pyOpenSci! Thank you @jsdodge for submitting THzTools and many thanks to @frank1010111 and @Romain-Peretti for reviewing this package! 😸

There are a few things left to do to wrap up this submission, @jsdodge:

- [ ] Activate [Zenodo](https://zenodo.org/), but specific make sure to **_watch_** the repo if you haven't already done so.
- [ ] Add the badge for pyOpenSci peer-review to the README.md of THzTools. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/209)`.

<IF JOSS SUBMISSION>
You're opting to submit this package to JOSS. Here are the next steps:

- [ ] @banesullivan: Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉
<IF JOSS SUBMISSION/>

These are for me, @banesullivan:

---
