### 2024-11-23 9b8d2ee
Add pyOpenSci badge; revise affiliations to Physics Department, Simon Fraser University
