### 2024-11-23 #209 @jsdodge
Thanks again, @banesullivan ! Please see my responses below.

> * [x]  Activate [Zenodo](https://zenodo.org/), but specificly make sure to **_watch_** the repo if you haven't already done so.

I assume you mean here that I have turned on GitHub integration for the repository? If so, then yes, that is on.

> * [x]  Tag and create a release to create a Zenodo version and DOI.
> * [x]  Add the badge for pyOpenSci peer-review to the README.md of THzTools. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/209)`.
> * [x]  Please fill out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7). All maintainers and reviewers should fill this out.
>
> You're opting to submit this package to JOSS. Here are the next steps:
> * [ ]  [@banesullivan](https://github.com/banesullivan): Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.
> * [x]  Login to the JOSS website and fill out the JOSS submission form using your Zenodo DOI. **When you fill out the form, be sure to mention and link to the approved pyOpenSci review.** JOSS will tag your package for expedited review if it is already pyOpenSci approved.

I didn't see a field for the Zenodo DOI, so I just included it in the comments. I tried entering it as the repository URL but it threw an error and insisted on a GitHub URL.

> * [ ]  Wait for a JOSS editor to approve the presubmission (which includes a scope check).
> * [ ]  Once the package is approved by JOSS, you will be given instructions by JOSS about updating the citation information in your README file.
> * [ ]  When the JOSS review is complete, add a comment to your review in the pyOpenSci software-review repo here that it has been approved by JOSS. An editor will then add the JOSS-approved label to this issue.
