+++ paper/paper.bib
@@ -94,7 +94,8 @@
   booktitle = {48th {{International Conference}} on {{Infrared}}, {{Millimeter}} and {{Terahertz Waves}} ({{IRMMW-THz}})},
   author = {Peretti, Romain},
   year = {2023},
-  address = {Montreal, Canada}
+  address = {Montreal, Canada},
+  doi = {10.1109/IRMMW-THz57677.2023.10299169}
 }
 
 @misc{peretti2024,
