### 2024-11-23 415daac
Add JOSS badge
