### 2024-11-24 #209 @jsdodge
Hi @banesullivan , when would be the deadline for providing pyOpenSci with a blog post?
