### 2024-11-26 #209 @banesullivan
Thought I'd drop a note that we will keep this issue open until THzTools finishes the review process with JOSS but THzTools is accepted by PyOpenSci at this point and the review has concluded 🎉
