### 2024-11-26 #209 @banesullivan
JOSS Pre-Review under way: https://github.com/openjournals/joss-reviews/issues/7521
