### 2024-12-02 #209 @banesullivan
JOSS review has been accepted 🎉  https://github.com/openjournals/joss-reviews/issues/7542
