### 2024-12-03 #209 @banesullivan
@jsdodge, should the "Software Package" citation in THzTool's README be updated to the JOSS citation? https://github.com/dodge-research-group/thztools?tab=readme-ov-file#software-package
