### 2024-12-03 41b6c1e
Update after JOSS acceptance
