### 2024-12-03 551b41b
Remove shell syntax highlighting from BibTeX citations in README.md
