### 2024-12-03 #209 @jsdodge
I updated the citation information and released v.0.5.5.post0. Thanks again!
