### 2025-03-26 #209 @banesullivan
Hey @jsdodge, it looks like you've moved THzTools to https://github.com/dodge-research-group/thztools

I'm updating the top level post with this link since we regularly scrape the metadata in these issues to build the package overviews on PyOpenSci's website: https://www.pyopensci.org/python-packages.html#accepted-pyopensci-packages
